requirements.txt
fastapi==0.108.0
pydantic==2.5.3
pydantic_core==2.14.6
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
transformers==4.40.1
uvicorn==0.25.0
sentencepiece==0.1.99
accelerate==0.29.3
bitsandbytes==0.41.2.post2
scipy==1.11.4
opencv-python==4.10.0.84
openai==1.42.0
easyocr==1.7.1
python-doctr==0.9.0
python-multipart